+++ src/main/java/net/thesquire/backroomsmod/block/custom/ElevatorButton.java
@@ -81,4 +81,5 @@ public class ElevatorButton extends ButtonBlock implements BlockEntityProvider {
 
         return super.onBreak(world, pos, state, player);
     }
+
 }